requirements.txt
httpx[http2]
discord.py
lmdb
PyJWT
mmh3
orjson
ff3